+++ examples/react-router/app/hello.js
@@ -1,0 +1,9 @@
+// Copyright Jon Williams 2017-2018. See LICENSE file.
+import React from 'react';
+
+export default () => (
+  <div>
+    <h2>Hello!</h2>
+    <p>This page only renders at the /hello route</p>
+  </div>
+);